SAR-4220 refactored and moved other views for story to same package
