+++ test/uk/gov/hmrc/vatsignupfrontend/views/principal/eligibility/ReturnDueSpec.scala
@@ -14,59 +14,40 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.vatsignupfrontend.views.principal
+package uk.gov.hmrc.vatsignupfrontend.views.principal.eligibility
 
-import org.jsoup.Jsoup
 import play.api.i18n.Messages.Implicits._
-import play.api.i18n.MessagesApi
 import play.api.test.FakeRequest
 import play.api.{Configuration, Environment}
-import uk.gov.hmrc.vatsignupfrontend.assets.MessageLookup.{MakingTaxDigitalSoftware => messages}
+import uk.gov.hmrc.vatsignupfrontend.assets.MessageLookup.{PrincipalReturnDue => messages}
 import uk.gov.hmrc.vatsignupfrontend.config.AppConfig
 import uk.gov.hmrc.vatsignupfrontend.views.ViewSpec
 
 
-class MakingTaxDigitalSoftwareViewSpec extends ViewSpec {
+class ReturnDueSpec extends ViewSpec {
 
   val env = Environment.simple()
   val configuration = Configuration.load(env)
 
-  val conf = new AppConfig(configuration, env)
-
-  lazy val messagesApi = app.injector.instanceOf[MessagesApi]
-
-  lazy val page = uk.gov.hmrc.vatsignupfrontend.views.html.principal.making_tax_digital_software(testCall)(
+  lazy val page = uk.gov.hmrc.vatsignupfrontend.views.html.principal.eligibility.return_due()(
     FakeRequest(),
     applicationMessages,
-    conf
+    new AppConfig(configuration, env)
   )
 
-  lazy val document = Jsoup.parse(page.body)
-
-  "Making Tax Digital Software View" should {
-
+  "The Return Due view" should {
     val testPage = TestView(
-      name = "Making Tax Digital View",
+      name = "Return Due View",
       title = messages.title,
       heading = messages.heading,
-      page = page,
-      haveSignOutInBanner = false
+      page = page
     )
-
     testPage.shouldHaveParaSeq(
-      messages.line1,
-      messages.line2
+      messages.line_1,
+      messages.line_2,
+      messages.line_3
     )
-
-    testPage.shouldHaveBulletSeq(
-      messages.bullet1,
-      messages.bullet2
-    )
-
-    testPage.shouldHaveALink(messages.linkId, messages.findSoftware, conf.softwareOptionsUrl)
-
-    testPage shouldHaveContinueButton()
+    testPage.shouldHaveALink(messages.link_id, messages.link_text, messages.link)
 
   }
-
 }
